Ajustes Caja Menor

+++ Spa.Domain.SpaEntities/AplicacionNomina.cs
@@ -8,7 +8,7 @@ namespace Spa.Domain.SpaEntities
         public List<Liquidacion> Empleados { get; set; }
         public DateTime Fecha_Nomina { get; set; }
         public string Id_Empresa { get; set; }
-        public float Total_Nomina { get; set; }
+        public decimal Total_Nomina { get; set; }
         public string Usuario_Registro { get; set; }
     }
 }
