+++ Spa.Domain.SpaEntities/EmpleadoServicio.cs
@@ -9,7 +9,7 @@ namespace Spa.Domain.SpaEntities
         public int Id_Empleado { get; set; }
         public string Id_Empresa_Servicio { get; set; }
         public int Id_Servicio { get; set; }
-        public float? Aplicacion_Nomina { get; set; }
+        public double? Aplicacion_Nomina { get; set; }
         public DateTime? Fecha_Creacion { get; set; }
         public string Usuario_Creacion { get; set; } = string.Empty;
         public DateTime? Fecha_Modificacion { get; set; }
